Move DataProviderUi to own folder
